# perf(idna): fast-path simple/ascii domains