Complements #760, together they reduce url-parsing benches by ~25-30% (or improve throughput by +33-40%)

Before (main 48fcbe1):
```
test long  ... bench:         484 ns/iter (+/- 6) = 88 MB/s
test short ... bench:         347 ns/iter (+/- 5) = 72 MB/s
```

Inter (#760):
```
test long  ... bench:         413 ns/iter (+/- 4) = 104 MB/s
test short ... bench:         298 ns/iter (+/- 9) = 83 MB/s
```

After (#760 + this #761)
```
test long  ... bench:         363 ns/iter (+/- 10) = 118 MB/s
test short ... bench:         247 ns/iter (+/- 12) = 101 MB/s
```